Extract out countdown formatter
