+++ go/app/javascript/activities_controller.js
@@ -1,3 +1,5 @@
+import CountdownFormatter from './countdown_formatter.js'
+
 export default class ActivitiesController extends Stimulus.Controller {
   static get targets () {
     return [ "current", "next", "remaining" ]
@@ -29,34 +31,3 @@ export default class ActivitiesController extends Stimulus.Controller {
   }
 }
 
-class CountdownFormatter {
-  format(targetTime) {
-    var remaining = this._diffNow(targetTime)
-    console.log(remaining.toObject())
-
-    var text = [ 
-      this._hoursString(remaining), 
-      this._minutesString(remaining),
-      this._secondsString(remaining)
-    ].filter((x) => x != null)
-
-    return text.join(', ').trim()
-  }
-
-  _diffNow(time) {
-    return luxon.DateTime.fromISO(time)
-      .diffNow(['hours', 'minutes', 'seconds', 'milliseconds'])
-  }
-
-  _hoursString(diff) { 
-    return diff.hours > 0 ? diff.hours + ' hours' : null
-  }
-
-  _minutesString(diff) { 
-    return diff.minutes > 0 ? diff.minutes + ' minutes' : null
-  }
-
-  _secondsString(diff) { 
-    return diff.seconds > 0 ? diff.seconds + ' seconds' : null
-  }
-}
